Shift the tokens into a separate, JWT package
Currently there is a little bit of confusing terminology. There are:

1. "Tokens" in the sense of OAuth, such an access or refresh token.
   These are opaque tokens to the spec¹
2. "JSON Web Tokens", which are commonly a format for the data chunk of
   an access token.

In addition, there's OpenID which includes a "non-standard" OAuth2 field
called "token_id", which *is* a JSON web token.

This commit tries to resolve this ambiguity by putting the JWTs in a
package called JWT.

—
1. https://www.oauth.com/oauth2-servers/access-tokens/
2. https://oauth.net/id-tokens-vs-access-tokens/
